+++ src/theme/components/helperErrorText.ts
@@ -3,13 +3,16 @@ interface HelperErrorTextBaseStyle {
 }
 
 const helperErrorText = {
+  parts: ["innerChild"],
   baseStyle: ({ isInvalid }: HelperErrorTextBaseStyle) => ({
-    marginTop: "xxs",
-    marginBottom: "0",
     fontSize: "desktop.helper.default",
     color: isInvalid ? "ui.error.primary" : null,
     _dark: {
       color: isInvalid ? "dark.ui.error.primary" : null,
+    },
+    innerChild: {
+      marginTop: "xxs",
+      marginBottom: "0",
     },
   }),
 };