+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates the background color of the "clear" button in Dark Mode for the `TextInput` component. This also affects the `SearchBar` component when a clear button is present in the `TextInput` component.
+
 ## 1.6.1 (June 22, 2023)
 
 ### Adds