+++ CHANGELOG.md
@@ -8,6 +8,15 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates the existing `heading` typography styles to add new default styles options for native heading elements, including `h1` thru `h6`.
+- Updates the existing `heading` typography styles to add new size options, including `display1` and `heading1` thru `heading6`.
+- Updates the existing `text` typography styles to add `overline` and `subtitle` options.
+- Updates the `Typography` docs to include the new options added to the `heading` and `text` typography styles.
+- Updates the `Design Tokens` docs to include the new tokens added to the `heading` and `text` typography styles.
+- Updates the `Heading` component to include `overline` and `subtitle` elements.
+
 ## 1.6.1 (June 22, 2023)
 
 ### Adds