+++ CHANGELOG.md
@@ -10,6 +10,12 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
+- Updates the existing `heading` typography styles to add new default styles options for native heading elements, including `h1` thru `h6`.
+- Updates the existing `heading` typography styles to add new size options, including `display1` and `heading1` thru `heading6`.
+- Updates the existing `text` typography styles to add `overline` and `subtitle` options.
+- Updates the `Typography` docs to include the new options added to the `heading` and `text` typography styles.
+- Updates the `Design Tokens` docs to include the new tokens added to the `heading` and `text` typography styles.
+- Updates the `Heading` component to include `overline` and `subtitle` elements.
 - Updates the spacing gap between main content in the `TemplateAppContainer` component.
 
 ## 1.6.1 (June 22, 2023)
@@ -21,6 +27,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the `"connect"` color variant to the `Breadcrumbs` component.
 - Adds internal `ChakraProvider` component to be able to conditionally include CSS styles through the `resetCSS` and `disableGlobalStyle` props.
 - Adds a `theme` prop to the `DSProvider` to merge in an additional theme file if necessary; this is to support to the `Header` and `Footer` components in the NYPL Header App.
+- Adds `Managing Focus` page to the `Accessibility Guide` in Storybook.
 
 ### Updates