+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `Design Tokens` docs to include the new tokens added to the `heading` and `text` typography styles.
 - Updates the `Heading` component to include `overline` and `subtitle` elements.
 - Updates the spacing gap between main content in the `TemplateAppContainer` component.
+- Updates the background color of the "clear" button in Dark Mode for the `TextInput` component. This also affects the `SearchBar` component when a clear button is present in the `TextInput` component.
 
 ## 1.6.1 (June 22, 2023)