+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `Heading` component to include `overline` and `subtitle` elements.
 - Updates the spacing gap between main content in the `TemplateAppContainer` component.
 - Updates the background color of the "clear" button in Dark Mode for the `TextInput` component. This also affects the `SearchBar` component when a clear button is present in the `TextInput` component.
+- Updates the spacing of the text in the `HelperErrorText` component.
+- Updates the `ComponentWrapper` component to conditionally render the `HelperErrorText` component.
 
 ## 1.6.1 (June 22, 2023)