+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the background color of the "clear" button in Dark Mode for the `TextInput` component. This also affects the `SearchBar` component when a clear button is present in the `TextInput` component.
 - Updates the spacing of the text in the `HelperErrorText` component.
 - Updates the `ComponentWrapper` component to conditionally render the `HelperErrorText` component.
+- Updates the version of the `react-datepicker` npm package.
 
 ## 1.6.1 (June 22, 2023)