+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the spacing of the text in the `HelperErrorText` component.
 - Updates the `ComponentWrapper` component to conditionally render the `HelperErrorText` component.
 - Updates the version of the `react-datepicker` npm package.
+- Updates the right-to-left visual display for the `Accordion` component.
 
 ## 1.6.1 (June 22, 2023)